Move ErasedLeaf to a submodule.
Move `ErasedLeaf` to a submodule of `erased`, so we can build several
different erased leaf and erased layer implementations in `erased`.
