+++ crates/dbsp/src/trace/layers/erased/erased_leaf/cursor.rs
